requirements.txt
streamlit
streamlit-folium
gpxpy
pillow
